Renamed vision.h files to vision_cpu.h and vision_cuda.h
