+++ torchvision/csrc/ROIAlign.h
@@ -1,9 +1,9 @@
 #pragma once
 
-#include "cpu/vision.h"
+#include "cpu/vision_cpu.h"
 
 #ifdef WITH_CUDA
-#include "cuda/vision.h"
+#include "cuda/vision_cuda.h"
 #endif
 
 // Interface for Python
