+++ torchvision/csrc/ROIPool.h
@@ -1,9 +1,9 @@
 #pragma once
 
-#include "cpu/vision.h"
+#include "cpu/vision_cpu.h"
 
 #ifdef WITH_CUDA
-#include "cuda/vision.h"
+#include "cuda/vision_cuda.h"
 #endif
 
 std::tuple<at::Tensor, at::Tensor> ROIPool_forward(
