+++ torchvision/csrc/cpu/nms_cpu.cpp
@@ -1,4 +1,4 @@
-#include "cpu/vision.h"
+#include "cpu/vision_cpu.h"
 
 template <typename scalar_t>
 at::Tensor nms_cpu_kernel(
